+++ server/src/main/java/org/opensearch/telemetry/TelemetrySettings.java
@@ -14,9 +14,6 @@ import org.opensearch.common.settings.Setting;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.common.unit.TimeValue;
 
-import java.util.List;
-import java.util.function.Function;
-
 /**
  * Wrapper class to encapsulate tracing related settings
  *
@@ -79,26 +76,11 @@ public class TelemetrySettings {
         Setting.Property.Dynamic
     );
 
-    /**
-     * Order of execution of samplers
-     */
-    public static final Setting<List<String>> TRACER_SPAN_SAMPLER_CLASSES = Setting.listSetting(
-        "telemetry.otel.tracer.span.sampler.classes",
-        List.of(
-            "org.opensearch.telemetry.tracing.sampler.ProbabilisticSampler",
-            "org.opensearch.telemetry.tracing.sampler.ProbabilisticTransportActionSampler"
-        ),
-        Function.identity(),
-        Setting.Property.NodeScope,
-        Setting.Property.Dynamic
-    );
-
     private volatile boolean tracingEnabled;
     private volatile double samplingProbability;
     private volatile double actionSamplingProbability;
     private final boolean tracingFeatureEnabled;
     private final boolean metricsFeatureEnabled;
-    private volatile List<String> samplersOrder;
 
     public TelemetrySettings(Settings settings, ClusterSettings clusterSettings) {
         this.tracingEnabled = TRACER_ENABLED_SETTING.get(settings);
@@ -106,20 +88,10 @@ public class TelemetrySettings {
         this.tracingFeatureEnabled = TRACER_FEATURE_ENABLED_SETTING.get(settings);
         this.metricsFeatureEnabled = METRICS_FEATURE_ENABLED_SETTING.get(settings);
         this.actionSamplingProbability = TRACER_SAMPLER_ACTION_PROBABILITY.get(settings);
-        this.samplersOrder = TRACER_SPAN_SAMPLER_CLASSES.get(settings);
 
         clusterSettings.addSettingsUpdateConsumer(TRACER_ENABLED_SETTING, this::setTracingEnabled);
         clusterSettings.addSettingsUpdateConsumer(TRACER_SAMPLER_PROBABILITY, this::setSamplingProbability);
         clusterSettings.addSettingsUpdateConsumer(TRACER_SAMPLER_ACTION_PROBABILITY, this::setActionSamplingProbability);
-        clusterSettings.addSettingsUpdateConsumer(TRACER_SPAN_SAMPLER_CLASSES, this::setSamplingOrder);
-    }
-
-    public void setSamplingOrder(List<String> order) {
-        this.samplersOrder = order;
-    }
-
-    public List<String> getSamplingOrder() {
-        return samplersOrder;
     }
 
     public void setTracingEnabled(boolean tracingEnabled) {